+++ docs/content/components/form/button/button-link.demo.tsx
@@ -1,0 +1,10 @@
+import { Button, Inline, Link } from '@marigold/components';
+
+export default () => (
+  <Inline space={2} alignY="center" alignX="center">
+    <Button variant="primary">Save</Button>
+    <Link variant="secondary" href="#">
+      Cancel
+    </Link>
+  </Inline>
+);