+++ themes/theme-docs/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @marigold/theme-docs
 
+## 2.1.0
+
+### Minor Changes
+
+- [#4054](https://github.com/marigold-ui/marigold/pull/4054) [`0fb763d`](https://github.com/marigold-ui/marigold/commit/0fb763ddd199c4f8f2477064d4008fdf22b949a4) Thanks [@sebald](https://github.com/sebald)! - feat: add dedicated export for tokens to all themes (`@marigold/<theme-name>/tokens`)
+
+### Patch Changes
+
+- Updated dependencies [[`5d53af4`](https://github.com/marigold-ui/marigold/commit/5d53af4ef32d8f70ae8d2d84db4fbfdd60998e79), [`965512c`](https://github.com/marigold-ui/marigold/commit/965512c113938cac629bb6cc518926f0d600b40f), [`9598df4`](https://github.com/marigold-ui/marigold/commit/9598df4ed6ac3fa72620d3b2b41d47a451a55d79)]:
+  - @marigold/components@9.0.1
+  - @marigold/system@9.0.1
+  - @marigold/theme-preset@1.3.21
+
 ## 2.0.5
 
 ### Patch Changes